feat : community mockdata revive, some logic refactoring
