+++ package.json
@@ -15,6 +15,7 @@
         "@tailwindcss/vite": "^4.1.18",
         "@tanstack/react-query": "^5.90.14",
         "axios": "^1.13.2",
+        "event-source-polyfill": "^1.0.31",
         "firebase": "^12.8.0",
         "lucide-react": "^0.562.0",
         "react": "^19.2.0",
