+++ src/pages/boards/free/free-detail-page/FreeDetailPage.jsx
@@ -1,17 +1,17 @@
-// import { useNavigate, useParams } from 'react-router-dom';
-// import { CommunityDetail } from '../../../../components/CommunityDetail';
+import { useNavigate, useParams } from 'react-router-dom';
+import { CommunityDetail } from '../../../../components/community/CommunityDetail';
 
-// export default function FreeDetailPage() {
-//     const { boardId, freeId } = useParams();
-//     const navigate = useNavigate();
+export default function FreeDetailPage() {
+    const { boardId, freeId } = useParams();
+    const navigate = useNavigate();
 
-//     const onNavigate = (key) => {
-//         if (key === 'community') navigate(`/boards/${boardId}/free`);
-//     };
+    const onNavigate = (key) => {
+        if (key === 'community') navigate(`/boards/${boardId}/free`);
+    };
 
-//     return (
-//         <>
-//             <CommunityDetail postId={freeId} onNavigate={onNavigate} />
-//         </>
-//     );
-// }
+    return (
+        <>
+            <CommunityDetail postId={freeId} onNavigate={onNavigate} />
+        </>
+    );
+}
